requirements.txt
gitpython>=3.0.8
pyinstaller>=3.6
